backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
django-cors-headers==4.3.*
psycopg2-binary==2.9.*
pymongo==4.6.*
numpy>=1.24.0
scikit-learn>=1.2.0
python-dotenv==1.0.*
requests>=2.31.0